refactor(repodb): simplify code structure
