+++ pkg/api/controller.go
@@ -24,10 +24,10 @@ import (
 	ext "zotregistry.io/zot/pkg/extensions"
 	"zotregistry.io/zot/pkg/extensions/monitoring"
 	"zotregistry.io/zot/pkg/log"
+	bolt "zotregistry.io/zot/pkg/meta/boltdb-wrapper"
+	dynamoParams "zotregistry.io/zot/pkg/meta/dynamodb-wrapper/params"
 	"zotregistry.io/zot/pkg/meta/repodb"
-	bolt "zotregistry.io/zot/pkg/meta/repodb/boltdb-wrapper"
-	dynamoParams "zotregistry.io/zot/pkg/meta/repodb/dynamodb-wrapper/params"
-	"zotregistry.io/zot/pkg/meta/repodb/repodbfactory"
+	"zotregistry.io/zot/pkg/meta/repodbfactory"
 	"zotregistry.io/zot/pkg/scheduler"
 	"zotregistry.io/zot/pkg/storage"
 	"zotregistry.io/zot/pkg/storage/cache"
